Update Jenkinsfile to use pipeline library
This updates the Jenkinsfile to use the helper functions defined in our
[pipeline library][1], as well as introducing some control flow
(explained below) to ensure that deployments happen as we expect.

[1]: https://github.com/hypothesis/pipeline-library

For more information on the `buildApp`/`releaseApp`/`deployApp` helper
functions, see the [pipeline library][1] documentation.

The other important change in this commit is the introduction of flow
control using `milestone` and `lock` pipeline steps. These are
introduced to ensure that deployments happen in-order and one at a time
(to each environment) when multiple builds occur in parallel on Jenkins.
They achieve this as follows:

The `milestone()` steps indicate points in the build where if a build
started later reaches a particular milestone before a build started
earlier, then the earlier build will be cancelled. To understand why
this is helpful, consider two builds executing the following pipeline:

    stage('test') {
        milestone()
        runTests()
    }

    stage('deploy') {
        milestone()
        deployApp()
    }

If build A and build B are started in order but run in parallel, then it
is possible for build B to get to the 'deploy' stage before build A.
(If, for example, build B happens to run on a more powerful node.) If
this happens, we may wish to ensure that build A's deployment doesn't
proceed and overwrite the build B's deployment.

The use of the `milestone()` step ensures this. When build A gets to the
second `milestone()`, Jenkins will notice that there exists a build that
started later which has already passed this milestone: build B. It will
then abort build A.

This doesn't solve all possible concurrency problems. While we have
prevented the possibility of an earlier build starting a stage after a
later build, it is still possible for build B to start a deployment
after build A's has started, but before it has finished. This is where
the `lock` step comes in.

We can use named locks to ensure that in parallel builds, only one build
can hold the lock. Consider the following pipeline:

    stage('test') {
        milestone()
        runTests()
    }

    stage('deploy') {
        lock(resource: 'deploy', inversePrecedence: true) {
             milestone()
             deployApp()
        }
    }

If build A reaches the 'deploy' stage first, it will take the lock, and
build B will wait. When build A finishes its deployment, build B will
proceed.

Conversely: if build B reaches the 'deploy' stage first, then it will
take the lock and build A will wait. When build B reaches the
`milestone()` step, it will detect the earlier job waiting on the lock
and terminate it.

There is one other important case to consider: when multiple builds
stack up waiting for the lock. Imagine that build C is currently holding
the lock and running a deployment. Before this deployment has finished,
builds D and E are started and both reach the 'deploy' stage where they
wait on the lock. The use of the `inversePrecedence` flag to the lock
step ensures that when build C finishes it is the *most recent build*
(i.e. build E in this case) which receives the lock and proceeds to a
deployment.

The only additional complication in our case is the desire for a manual
gating stage before a production deployment. This is achieved using the
`input` step, and it is important to know that just as a `milestone()`
can cancel earlier builds waiting on a lock, it can also cancel earlier
builds waiting on `input`.

For more on `stage`, `milestone`, and `lock`, see [this Jenkins blog
post on the topic][2].

[2]: https://jenkins.io/blog/2016/10/16/stage-lock-milestone/