Further reduce the annotation reindexing rate
We're still seeing timeouts from Elasticsearch even when trying to bulk
index only 1000 annotations at a time. Reduce it to 400 to see if that
clears it. I don't think we can afford to go much lower than this or the
reindexing task will be in danger of being unable to keep up with normal
traffic (which regularly spikes up to around 200 new annotations per
minute).