Replicate the tox-pip-sync removal experiment from `h`
Some work was done to test the effects of removing `tox-pip-sync` from
our tox workflow in `h`. But `h` is one of the slowest performing apps
where any changes are the most likely to be swallowed up by the main
time taken to run the actions.

H-periodic offers a much better test bed to show this work in a more
honest light as the operations here are much faster and so the
overheads are more relevant.