Add GitHub environment URLs
Since h-periodic doesn't have a public URL or any other URL to go to in
order to test it, use the URLs of h-periodic's QA, production and Canada
Papertrail groups for the environment URLs. This means you can click on
the environment URL in GitHub and be taken straight to the Papertrail
logs.