Reduce LMS email digests batch size to 50
The previous value of 1000 was just an arbitrary number chosen by me and I think it's probably too large: it could take the task a long time to send 1000 emails.

In future we may split the actual email sending out into a separate task which sends one email per task instance. If we do that then we may be able to increase the batch size again.