Decrease send_instructor_email_digest_tasks() batch size
Decrease the `batch_size` of the `send_instructor_email_digest_tasks()`
Celery task.

This is because I want to try decreasing the task's `batch_size` and
correspondingly increasing its `rate_limit` (see
https://github.com/hypothesis/lms/pull/5827) to see what effect that has
on performance, with an eye towards potentially removing the batching
feature altogether in order to simplify adding future features to this
code.