Upgrade python dependencies
This is the result of running:

`make requirements --always-make args='--upgrade'`

Non top level dependencies can become outdated as dependabot doesn't
send PR for those.

On top of that upgrading single packages with all the existing
constraints might make impassible to satisfy all constraints.

This approach makes sure we are up to date with all our dependencies as
specified on the .in files but without considering any constraints in
the .txt files.